chore(jpaserver): Configure JPA server
Configure the HAPI JPA server for the current needs. The config
contains the basic needs and more to storage and retrieve
resoruces from the FHIR database (postgres). Once deemed safe
extra logic or components may be removed.

A UUID is generated for newly created resoruces as the fhir_id
column in the database. In addition to creating resoruces,
referenced resource in the newly created resource should also
be modified to include new UUID to ensure the UUID match the
referened resource and allow for proper ID searching.

Example Observation --> Encounter connection
- get Patient resource from external repo
- create new Patient resource in FHIR server
- get Observation resource from external repo
- create new Observation resources using new Patient UUID
- update Observation resource refs with new UUID (e.g, Encounter)
- get Encounter resources from external repo
- create new Encounter resources using new Patient UUID
- update Encounter resource refs with new UUID